scraper/pcpartpicker/scraper/pyproject.toml
[project]
name = "pcpartpicker-scraper"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "firecrawl>=4.12.0",
    "python-dotenv>=1.2.1",
]
